GBC.emu and misc updates/fixes
* Imagine: Delete GLDrawableHolder's drawable in onExit handler since some EGL implementations expect all resources deleted before calling eglTerminate()
* EmuFramework: Add EmuApp::defaultRenderPixelFormat()
* EmuFramework: Add EmuSystem::prepareVideo() and clear EmuVideo image afterwards
* EmuFramework: Update EmuVideoLayer effects parameters so the pixel format is passed along with the effect type
* EmuFramework: Update "Effect Color Format" option auto behavior so it matches the emulated system's render format
* EmuFramework: Call EmuVideoLayer::setOverlay() whenever EmuVideo's format changes to keep its line value correct
* EmuFramework: Don't pass inactive EmuAudio to EmuSystemTask::runFrame()
* GBC.emu: Update core to Gambatte GIT 56e3371 (2020.03.13)
* GBC.emu: Use Gambatte's internal RGBA8888 rendering and add option to down-sample to RGB565 for older devices
* GBC.emu: Retain the framebuffer to save screenshots into save-states for future use
* GBC.emu: Improve unchanged frame handling to prevent audio/video hiccups

+++ GBC.emu/src/libgambatte/src/video/next_m0_time.h
@@ -8,10 +8,10 @@ public:
 	NextM0Time() : predictedNextM0Time_(0) {}
 	void predictNextM0Time(class PPU const &v);
 	void invalidatePredictedNextM0Time() { predictedNextM0Time_ = 0; }
-	unsigned predictedNextM0Time() const { return predictedNextM0Time_; }
+	unsigned long predictedNextM0Time() const { return predictedNextM0Time_; }
 
 private:
-	unsigned predictedNextM0Time_;
+	unsigned long predictedNextM0Time_;
 };
 
 }
